refactor colbert imports
